Move BinaryTableFilter to proper package
