+++ src/main/java/io/stargate/sgv2/jsonapi/service/operation/filters/table/BinaryTableFilter.java
@@ -1,7 +1,6 @@
-package io.stargate.sgv2.jsonapi.service.operation.filters.table.codecs;
+package io.stargate.sgv2.jsonapi.service.operation.filters.table;
 
 import io.stargate.sgv2.jsonapi.api.model.command.clause.filter.EJSONWrapper;
-import io.stargate.sgv2.jsonapi.service.operation.filters.table.NativeTypeTableFilter;
 
 /** Filter to use any JSON Binary (EJSON/Base64-encoded) against a table column. */
 public class BinaryTableFilter extends NativeTypeTableFilter<EJSONWrapper> {
